show machine type in review card
